moved frontend test files to the frontend directory
